vm_tools: concierge: migrate the file format for vmm-swap tbw policy
Use TbwHistoryEntryContainer to serialize repeated entries as
VmmSwapUsagePolicy do.

Since the old/new file format are incompatible, we will use a new file
/var/lib/vm_concierge/vmm_swap_policy/tbw_history2 and migrate data if
there is the old file.

We don't need to update concierge minijail settings or ChromeOS settings
because we set the tbw history file on directory basis.

BUG=b:289461315
TEST=cros_run_unit_tests --board=$BOARD --packages
chromeos-base/vm_host_tools

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4659521
Commit-Queue: Shin Kawamura <kawasin@google.com>
Tested-by: Shin Kawamura <kawasin@google.com>

+++ vm_tools/concierge/vmm_swap_tbw_policy.cc
@@ -16,6 +16,7 @@
 #include <base/files/file_path.h>
 #include <base/files/file_util.h>
 #include <base/logging.h>
+#include <base/strings/strcat.h>
 #include <base/strings/string_number_conversions.h>
 #include <base/time/time.h>
 #include <brillo/files/file_util.h>
@@ -26,24 +27,21 @@
 namespace vm_tools::concierge {
 
 namespace {
-bool WriteEntry(base::File& file, uint64_t bytes_written, base::Time time) {
-  TbwHistoryEntry entry;
-  uint8_t message_size_buf[1];
-  entry.set_time_us(time.ToDeltaSinceWindowsEpoch().InMicroseconds());
-  entry.set_size(bytes_written);
-  // TbwHistoryEntry message is less than 127 bytes. The MSB is reserved for
-  // future extensibility.
-  if (entry.ByteSizeLong() > 127) {
-    LOG(ERROR) << "Tbw history message is bigger than 127 bytes: "
-               << entry.ByteSizeLong() << " bytes";
-    return false;
+constexpr char kOldHistoryFileName[] = "tbw_history";
+base::expected<size_t, std::string> WriteEntry(base::File& file,
+                                               uint64_t bytes_written,
+                                               base::Time time) {
+  // Consecutively serialized bytes from multiple TbwHistoryEntryContainers
+  // can be deserialized as single merged TbwHistoryEntryContainer.
+  TbwHistoryEntryContainer container;
+  TbwHistoryEntry* entry = container.add_entries();
+  entry->set_time_us(time.ToDeltaSinceWindowsEpoch().InMicroseconds());
+  entry->set_size(bytes_written);
+  if (container.SerializeToFileDescriptor(file.GetPlatformFile())) {
+    return base::ok(container.GetCachedSize());
+  } else {
+    return base::unexpected("failed to write tbw history");
   }
-  message_size_buf[0] = entry.ByteSizeLong();
-  if (!base::WriteFileDescriptor(file.GetPlatformFile(), message_size_buf)) {
-    LOG(ERROR) << "Failed to write tbw history entry size";
-    return false;
-  }
-  return entry.SerializeToFileDescriptor(file.GetPlatformFile());
 }
 }  // namespace
 
@@ -72,6 +70,33 @@ bool VmmSwapTbwPolicy::Init(base::FilePath path, base::Time now) {
   }
   history_file_path_ = path;
 
+  // The old format file was named "tbw_history". If there is the file, load the
+  // history from the old file and recreate a new file from the history.
+  // TODO(b/289975202): Remove this 2 milestones (M119) after since we care
+  // about the last 28 days history.
+  base::FilePath old_file_path =
+      history_file_path_.DirName().Append(kOldHistoryFileName);
+  if (base::PathExists(old_file_path)) {
+    LOG(INFO) << "Old tbw history file is found. Recreate a new history file.";
+    base::File old_file = base::File(
+        old_file_path, base::File::FLAG_OPEN | base::File::FLAG_READ);
+    // Remove the old file whether it succeeds to load history or not.
+    brillo::DeleteFile(old_file_path);
+    if (old_file.IsValid()) {
+      auto result = LoadFromOldFormattedFile(old_file, now);
+      if (result.has_value()) {
+        // The file was old formatted. Replace the old history file with the new
+        // file format.
+        return RotateHistoryFile(now);
+      } else {
+        LOG(ERROR) << "Failed to load old tbw history file: " << result.error();
+      }
+    } else {
+      LOG(ERROR) << "Failed to open old tbw history file: "
+                 << old_file.ErrorToString(old_file.error_details());
+    }
+  }
+
   history_file_ =
       base::File(path, base::File::FLAG_CREATE | base::File::FLAG_READ |
                            base::File::FLAG_WRITE);
@@ -92,12 +117,12 @@ bool VmmSwapTbwPolicy::Init(base::FilePath path, base::Time now) {
     history_file_ = base::File(history_file_path_, base::File::FLAG_OPEN |
                                                        base::File::FLAG_READ |
                                                        base::File::FLAG_WRITE);
-    auto num_entries = LoadFromFile(history_file_, now);
-    if (num_entries.has_value()) {
-      entries_in_file_ = num_entries.value();
+    auto file_size = LoadFromFile(history_file_, now);
+    if (file_size.has_value()) {
+      history_file_size_ = file_size.value();
       return true;
     } else {
-      LOG(ERROR) << "Failed to load tbw history file: " << num_entries.error();
+      LOG(ERROR) << "Failed to load tbw history file: " << file_size.error();
     }
   } else {
     LOG(ERROR) << "Failed to create tbw history file: "
@@ -119,16 +144,16 @@ void VmmSwapTbwPolicy::Record(uint64_t bytes_written, base::Time time) {
   AppendEntry(bytes_written, time);
 
   if (history_file_.IsValid()) {
-    if (entries_in_file_ >= kMaxEntriesInFile) {
-      if (!RotateHistoryFile(time)) {
-        LOG(ERROR) << "Failed to rotate tbw file";
-        // Stop writing a new entry to the history file.
-        DeleteFile();
-        return;
-      }
+    if (history_file_size_ >= kMaxFileSize - kMaxEntrySize &&
+        !RotateHistoryFile(time)) {
+      LOG(ERROR) << "Failed to rotate tbw file";
+      // Stop writing a new entry to the history file.
+      DeleteFile();
+      return;
     }
-    if (WriteEntry(history_file_, bytes_written, time)) {
-      entries_in_file_++;
+    auto entry_size = WriteEntry(history_file_, bytes_written, time);
+    if (entry_size.has_value()) {
+      history_file_size_ += entry_size.value();
     } else {
       LOG(ERROR) << "Failed to write tbw entry to file";
       // Delete the history file since the file content is now broken.
@@ -165,9 +190,50 @@ base::expected<size_t, std::string> VmmSwapTbwPolicy::LoadFromFile(
     base::File& file, base::Time now) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   if (!file.IsValid()) {
-    return base::unexpected("usage history file is invalid to load");
+    return base::unexpected("tbw history file is invalid to load");
   }
-  int entries_in_file = 0;
+
+  int64_t file_size = file.GetLength();
+  if (file_size < 0) {
+    return base::unexpected("get length of history file: " +
+                            file.ErrorToString(file.GetLastFileError()));
+  } else if (file_size > kMaxFileSize) {
+    // Validates the file size because this loads all entries at once.
+    return base::unexpected(
+        base::StrCat({"tbw history file: ", base::NumberToString(file_size),
+                      " is bigger than ", base::NumberToString(kMaxFileSize)}));
+  }
+
+  TbwHistoryEntryContainer container;
+  if (!container.ParseFromFileDescriptor(file.GetPlatformFile())) {
+    return base::unexpected("parse tbw history");
+  }
+
+  base::Time previous_time;
+  for (auto entry : container.entries()) {
+    base::Time time = base::Time::FromDeltaSinceWindowsEpoch(
+        base::Microseconds(entry.time_us()));
+    if ((now - time).is_negative()) {
+      return base::unexpected("tbw history file has invalid time (too new)");
+    } else if ((time - previous_time).is_negative()) {
+      return base::unexpected(
+          "tbw history file has invalid time (old than lastest)");
+    }
+    AppendEntry(entry.size(), time);
+
+    previous_time = time;
+  }
+
+  return file_size;
+}
+
+base::expected<void, std::string> VmmSwapTbwPolicy::LoadFromOldFormattedFile(
+    base::File& file, base::Time now) {
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (!file.IsValid()) {
+    return base::unexpected("tbw history file is invalid to load");
+  }
+
   google::protobuf::io::FileInputStream input_stream(file.GetPlatformFile());
   TbwHistoryEntry entry;
   base::Time previous_time;
@@ -205,10 +271,9 @@ base::expected<size_t, std::string> VmmSwapTbwPolicy::LoadFromFile(
           "tbw history file has invalid time (old than lastest)");
     }
     AppendEntry(entry.size(), time);
-    entries_in_file++;
     previous_time = time;
   }
-  return entries_in_file;
+  return base::ok();
 }
 
 void VmmSwapTbwPolicy::AppendEntry(uint64_t bytes_written, base::Time time) {
@@ -236,14 +301,16 @@ bool VmmSwapTbwPolicy::RotateHistoryFile(base::Time time) {
     return false;
   }
 
-  entries_in_file_ = 0;
+  history_file_size_ = 0;
   for (auto iter = tbw_history_.Begin(); iter; ++iter) {
     if ((time - iter->started_at) < base::Days(28)) {
-      if (WriteEntry(tmp_file, iter->size, iter->started_at)) {
+      auto entry_size = WriteEntry(tmp_file, iter->size, iter->started_at);
+      if (entry_size.has_value()) {
+        history_file_size_ += entry_size.value();
+      } else {
         LOG(ERROR) << "Failed to write entries to new tbw history file";
         return false;
       }
-      entries_in_file_++;
     }
   }
 
@@ -271,7 +338,7 @@ void VmmSwapTbwPolicy::DeleteFile() {
   }
   // Stop writing entries to the file.
   history_file_.Close();
-  entries_in_file_ = 0;
+  history_file_size_ = 0;
 }
 
 }  // namespace vm_tools::concierge
